requirements.txt
# Pipoo Desktop Assistant - Requirements
# Python 3.8+ required

# Speech Recognition
SpeechRecognition==3.10.0
pyaudio==0.2.14
# Text-to-Speech
pyttsx3==2.90
# API Requests
requests==2.31.0
#Environment Variables
python-dotenv==1.0.0
# GUI (comes with Python, but listing for clarity)
# tkinter - built-in with Python
# Additional utilities
Pillow==10.1.0
# Optional but recommended
colorama==0.4.6